fix: Inset rename
